ca4471b Add missing header definition in UI file
30c44f9 Fix qt module dependency
8d0c90d Fix property type cast to properly display remaining time
1dc3fec Fix 13439: Finer control over Lock Interactive Render
b97ee0e Fix interactive render delay after modularisation

+++ ParaViewCore/ClientServerCore/Rendering/vtkPVRenderView.h
@@ -363,7 +363,7 @@ public:
   //*****************************************************************
   // Forward to vtkPVGenericRenderWindowInteractor.
   void SetCenterOfRotation(double x, double y, double z);
-  void SetNonInteractiveRenderDelay(unsigned int seconds);
+  void SetNonInteractiveRenderDelay(double seconds);
 
   //*****************************************************************
   // Forward to vtkLightKit.
